refactor: move ampd-sdk to its own crate
