+++ packages/ampd-proto/build.rs
@@ -2,7 +2,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     tonic_build::configure()
         .build_server(false)
         .build_client(true)
-        .compile_protos(&["ampd-proto/ampd/v1/ampd.proto"], &["ampd-proto/ampd/v1"])?;
+        .compile_protos(
+            &["proto-files/ampd/v1/ampd.proto"],
+            &["proto-files/ampd/v1"],
+        )?;
 
     Ok(())
 }
