+++ packages/ampd-proto/src/lib.rs
@@ -1,0 +1,5 @@
+pub use proto::*;
+
+mod proto {
+    tonic::include_proto!("ampd.v1");
+}
