+++ packages/ampd-sdk/src/grpc/mod.rs
@@ -1,5 +1,1 @@
-pub mod proto {
-    tonic::include_proto!("ampd.v1");
-}
-
 pub mod client;
