+++ packages/events/Cargo.toml
@@ -6,6 +6,7 @@ edition = { workspace = true }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+ampd-proto = { workspace = true }
 axelar-wasm-std = { workspace = true }
 base64 = "0.21.2"
 cosmrs = { version = "0.14.0", features = ["cosmwasm"] }
